pyproject.toml
# https://python-poetry.org/docs/pyproject/

[tool.poetry]
name = "flask_web_api_template"
version = "0.1.0"
description = ""
authors = ["Tatsuru Kawatsu <t-kawatsu@jogoj.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1"
PyYAML = "^5.4.1"
flask-restx = "^0.1.1"
python-dotenv = "^0.17"
blinker = "^1.4"
Flask-Caching = "^1.0"

[tool.poetry.dev-dependencies]
flake8 = "3.9.*"
pytest = "^6.2.3"
pytest-cov = "^2.11.1"

[tool.poetry.scripts]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-vx"
testpaths = [
    "tests"
]
python_files = "test_*.py"
# durations = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
